Allow custom request and response model factories
Sometimes, it's needed custom factories for request and
response models to handle incoming request in the
middleware. You could solve this by extending the
middleware but probably end up copying a lot of code.
With this development, you can configure custom classes
to solve this.

`SILKY_REQUEST_MODEL_FACTORY_CLASS` and
`SILKY_RESPONSE_MODEL_FACTORY_CLASS` are settings that
allow customization of factories.